Revert "feat(cli): autostart without a console window"
This reverts commit 08494b46ddc3b9db4ea9b7b1c3f7f007af9c9087.
