--- komorebic/src/bin/komorebic-no-console.rs
@@ -1,5 +1,0 @@
-#![windows_subsystem = "windows"]
-
-fn main() -> color_eyre::Result<()> {
-    komorebic::main()
-}
